Add more Admin 'friendly' functionality 1. shop.actions   - Add 'history' & 'location' to 'addCategory' action to push Admin back to /shop   - Dispatch 'setAlert' in 'addCategory' & 'deleteCategory' to let Admin aware of successful or failed actions 2. Category.component   - Set 'name' as the Category being viewed 3. DashboardBtns.component   - Add 'window.confirm' before dispatching 'deleteCategory' 4. AddCategory.component   - Add 'history' & 'location' to 'addCategory' action

+++ client/src/components/dashboard-btns/DashboardBtns.component.js
@@ -18,8 +18,14 @@ const DashboardBtns = ({
 
   /**  DELETE ACTION */
   const handleDelete = e => {
-    deleteCategory(remove);
-    history.push(`${location.state.from}`);
+    if (
+      window.confirm(
+        `Are you sure you want to delete ${name}? This cannot be undone.`
+      )
+    ) {
+      deleteCategory(remove);
+      history.push(`${location.state.from}`);
+    }
   };
 
   return (
